:package: cli: Transfered issue variables to .gitgo
