+++ lib/inquirer.js
@@ -1,49 +1,8 @@
 var inquirer = require('inquirer');
 var emoji = require('node-emoji')
 const { getIssue } = require('./issue.js');
-const fs = require('fs');
-
-/*
-Sample Usage
-
-1. Reading content
-jsonReader('./.gitgo', (err, conf) => {
-    if (err) {
-        console.log(err)
-        return
-    }
-    console.log(conf)
-});
-
-2. Updating conent
-jsonReader('./.gitgo', (err, conf) => {
-    if (err) {
-        console.log('Error reading file:',err)
-        return
-    }
-    conf.commit_guidelines[0] = "somethingelse:"
-    conf.emojis.config = "thought_balloon"
-    fs.writeFile('./.gitgo', JSON.stringify(conf), (err) => {
-            if (err) console.log('Error writing file:', err)
-    })
-})
-*/
-function jsonReader(filePath, cb) {
-    fs.readFile(filePath, (err, fileData) => {
-        if (err) {
-            return cb && cb(err)
-        }
-        try {
-            const object = JSON.parse(fileData)
-            return cb && cb(null, object)
-        } catch(err) {
-            return cb && cb(err)
-        }
-    })
-}
 
 module.exports = {
-    
     getQuestions: async () => {
 
         inquirer.prompt([
